% Part of the R eegkit package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{eegkit-internal}

%% List of Internal Functions Called
\alias{colorbar3d}
\alias{trimhead}
\alias{voltcol}


\title{Internal functions for eegkit package}
\description{
  Internal functions for eegkit package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

