% Part of the R eegkit package, 
% Nathaniel E. Helwig <helwig@umn.edu>

\name{eegkit-internal}

%% List of Internal Functions Called
\alias{colorbar3d}
\alias{trimhead}
\alias{voltcol}
\alias{p1t}
\alias{n1t}
\alias{p2t}
\alias{n2t}
\alias{p3t}
\alias{p1s}
\alias{n1s}
\alias{p2s}
\alias{n2s}
\alias{p3s}


\title{Internal Functions for eegkit Package}
\description{
  Internal functions for eegkit package.
}
\details{
  These functions are not to be called by the user.
}
\keyword{ internal }

